Renaming package and other required files to reflect new package name; Linking new package to example
