+++ example/package.json
@@ -20,6 +20,6 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "4.9.4",
-    "use-cardano": "link:.."
+    "mynth-use-cardano": "link:.."
   }
 }
