+++ example/pages/examples/transaction.tsx
@@ -1,8 +1,8 @@
 import { ExampleWrapper } from "components/ExampleWrapper"
-import { options } from "config/use-cardano-options"
+import { options } from "config/mynth-use-cardano-options"
 import { useTransaction } from "hooks/use-transaction"
 import styles from "styles/example.module.css"
-import { CardanoWalletSelector, useCardano } from "use-cardano"
+import { CardanoWalletSelector, useCardano } from "mynth-use-cardano"
 
 const TransactionExample = () => {
   const { isValid, lucid, networkId } = useCardano()
