+++ example/pages/examples/warnings-and-errors.tsx
@@ -1,7 +1,7 @@
 import { ExampleWrapper } from "components/ExampleWrapper"
-import { options } from "config/use-cardano-options"
+import { options } from "config/mynth-use-cardano-options"
 import styles from "styles/example.module.css"
-import { CardanoWalletSelector, useCardano } from "use-cardano"
+import { CardanoWalletSelector, useCardano } from "mynth-use-cardano"
 
 const WarningsAndErrorsExample = () => {
   const { walletApiError, accountError, networkWarning, accountWarning } = useCardano()
