+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "use-cardano",
+  "name": "mynth-use-cardano",
   "type": "module",
   "version": "1.1.0",
   "engines": {
@@ -8,7 +8,7 @@
   "description": "A react context, hook, and set of components that makes interacting with the Cardano blockchain easy",
   "main": "dist/index.js",
   "module": "dist/index.js",
-  "types": "src/types/use-cardano.d.ts",
+  "types": "src/types/mynth-use-cardano.d.ts",
   "keywords": [
     "cardano",
     "dApp",
@@ -19,11 +19,6 @@
   ],
   "author": "Alan Smithee (ggnore.alan.smithee@gmail.com)",
   "license": "MIT",
-  "homepage": "https://use-cardano.alangaming.com",
-  "repository": "use-cardano/use-cardano",
-  "bugs": {
-    "url": "https://www.github.com/use-cardano/use-cardano/issues"
-  },
   "scripts": {
     "build": "rollup -c",
     "dev:package": "rollup -c -w &",
