+++ src/hooks/use-cardano-initialization.ts
@@ -1,6 +1,6 @@
 import { useLucidAndWalletApi } from "hooks/use-lucid-and-wallet-api"
 import { useWalletProviders } from "hooks/use-wallet-providers"
-import { UseCardanoOptionsWithDefaults } from "use-cardano"
+import { UseCardanoOptionsWithDefaults } from "mynth-use-cardano"
 
 export const useCardanoInitialization = (options: UseCardanoOptionsWithDefaults) => {
   useWalletProviders(options)
