+++ src/lib/supported-wallet-providers.ts
@@ -1,3 +1,3 @@
-import { WalletProvider } from "use-cardano"
+import { WalletProvider } from "mynth-use-cardano"
 
 export const supportedWalletProviders: WalletProvider[] = ["nami", "eternl", "gerowallet", "flint"]
